代码调整

+++ gms-server/src/main/java/org/gms/dao/entity/PlayerdiseasesDO.java
@@ -38,6 +38,6 @@ public class PlayerdiseasesDO implements Serializable {
 
     private Integer mobskilllv;
 
-    private Integer length;
+    private Long length;
 
 }
